# [Nim] adapt approaches in nim_con to to single-threaded impl and improve perf in both