--- a/nim/related.nimble
+++ b/nim/related.nimble
@@ -2,7 +2,7 @@
 
 version       = "0.1.0"
 author        = "jinyus"
-description   = "A new awesome nimble package"
+description   = "Nim implementation for related_post_gen"
 license       = "MIT"
 srcDir        = "src"
 bin           = @["related"]
@@ -11,7 +11,7 @@ bin           = @["related"]
 # Dependencies
 
 requires "nim >= 2.0.0"
-requires "jsony"
+requires "jsony#head"
 requires "xxhash"
 
 const STAT_STEPS = 10