--- a/nim_con/config.nims
+++ b/nim_con/config.nims
@@ -10,9 +10,8 @@ switch("nimcache", cacheSubdir)
 
 --cc:clang
 --outdir:build
---threads:on
---tlsEmulation:off
---warning:"Effect:off"
+--tlsEmulation:off # default on|off varies by platform
+--warning:"Effect:off" # suppress noisy compiler warnings re: malebolgia
 
 when defined(profileGen):
   --hints:off