As I mentioned in #300, I'm trying things out on a free tier Azure VM with the same specs as your benchmarking VM.

With changes in this PR, I'm seeing some speedups
```
Nim Concurrent:

        Benchmark 1: ./build/related_con
        Processing time (w/o IO): 23.264ms
        Processing time (w/o IO): 22.754ms
        Processing time (w/o IO): 22.866ms
        Processing time (w/o IO): 22.699ms
        Processing time (w/o IO): 22.644ms
        Processing time (w/o IO): 22.95ms
        Processing time (w/o IO): 22.575ms
        Processing time (w/o IO): 22.785ms
        Processing time (w/o IO): 22.73ms
        Processing time (w/o IO): 22.623ms
        Processing time (w/o IO): 22.603ms
        Processing time (w/o IO): 22.645ms
        Processing time (w/o IO): 22.583ms
          Time (mean ± σ):      45.2 ms ±   0.8 ms    [User: 77.1 ms, System: 7.0 ms]
          Range (min … max):    43.8 ms …  46.6 ms    10 runs
         
Nim Concurrent:

        Benchmark 1: ./build/related_con
        Processing time (w/o IO): 309.609ms
        Processing time (w/o IO): 313.802ms
        Processing time (w/o IO): 314.481ms
          Time (mean ± σ):     411.4 ms ±   6.8 ms    [User: 955.1 ms, System: 42.3 ms]
          Range (min … max):   406.6 ms … 416.3 ms    2 runs
         
Nim Concurrent:

        Benchmark 1: ./build/related_con
        Processing time (w/o IO): 3214.735ms
        Processing time (w/o IO): 2883.433ms
        Processing time (w/o IO): 2883.317ms
          Time (mean ± σ):      3.164 s ±  0.001 s    [User: 8.772 s, System: 0.078 s]
          Range (min … max):    3.164 s …  3.165 s    2 runs
          
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 27.721ms
        Processing time (w/o IO): 27.495ms
        Processing time (w/o IO): 27.446ms
        Processing time (w/o IO): 27.396ms
        Processing time (w/o IO): 27.313ms
        Processing time (w/o IO): 27.65ms
        Processing time (w/o IO): 27.512ms
        Processing time (w/o IO): 27.363ms
        Processing time (w/o IO): 27.346ms
        Processing time (w/o IO): 27.336ms
        Processing time (w/o IO): 27.423ms
        Processing time (w/o IO): 27.309ms
        Processing time (w/o IO): 27.504ms
          Time (mean ± σ):      49.6 ms ±   0.5 ms    [User: 40.1 ms, System: 9.5 ms]
          Range (min … max):    49.1 ms …  50.9 ms    10 runs
         
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 379.217ms
        Processing time (w/o IO): 380.812ms
        Processing time (w/o IO): 379.355ms
          Time (mean ± σ):     467.8 ms ±   0.2 ms    [User: 435.8 ms, System: 31.9 ms]
          Range (min … max):   467.6 ms … 468.0 ms    2 runs
         
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 3400.621ms
        Processing time (w/o IO): 3398.67ms
        Processing time (w/o IO): 3399.28ms
          Time (mean ± σ):      3.680 s ±  0.002 s    [User: 3.599 s, System: 0.076 s]
          Range (min … max):    3.678 s …  3.681 s    2 runs
```
There seems to be a lot of runtime variability in the VM, probably because the box it's running on is more/less busy at various times.

So I'm not confident you'll get improved numbers, i.e. it might flip around again with `nim_con` being slower than `nim`, but let's see how it goes.

I've tried to explain clearly the why/what in each  commit message. I didn't completely overhaul the structure and code of `nim`, just adapted some of what I did in `nim_con` where it seemed like it should speedup the single-threaded impl.

I have some ideas about switching to a different task manager library for `nim_con`. There are several promising options besides `malebolgia` and they each take different approaches. Maybe one of those approaches will be better suited to the Azure VM environment, will be interesting to see how it goes.