@jinyus 2023-10-18
Ran 3 times with the same results:

```
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 31.314ms
        Processing time (w/o IO): 31.293ms
        Processing time (w/o IO): 30.912ms
        Processing time (w/o IO): 31.081ms
        Processing time (w/o IO): 31.201ms
        Processing time (w/o IO): 31.258ms
        Processing time (w/o IO): 31.206ms
        Processing time (w/o IO): 31.268ms
        Processing time (w/o IO): 30.987ms
        Processing time (w/o IO): 31.117ms
        Processing time (w/o IO): 30.878ms
        Processing time (w/o IO): 31.215ms
        Processing time (w/o IO): 31.067ms
          Time (mean ± σ):      57.3 ms ±   0.3 ms    [User: 48.6 ms, System: 8.7 ms]
          Range (min … max):    56.6 ms …  57.6 ms    10 runs
         
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 438.542ms
        Processing time (w/o IO): 439.15ms
        Processing time (w/o IO): 438.326ms
          Time (mean ± σ):     541.8 ms ±   0.9 ms    [User: 498.0 ms, System: 43.7 ms]
          Range (min … max):   541.2 ms … 542.5 ms    2 runs
         
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 3867.711ms
        Processing time (w/o IO): 3871.571ms
        Processing time (w/o IO): 3863.995ms
          Time (mean ± σ):      4.236 s ±  0.027 s    [User: 4.127 s, System: 0.093 s]
          Range (min … max):    4.217 s …  4.255 s    2 runs
```