@jinyus 2023-10-18
Fresh run from main branch:

`sudo docker run -it --rm -e TEST_NAME=nim databench`

```
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 30.636ms
        Processing time (w/o IO): 30.399ms
        Processing time (w/o IO): 30.472ms
        Processing time (w/o IO): 30.361ms
        Processing time (w/o IO): 30.541ms
        Processing time (w/o IO): 30.471ms
        Processing time (w/o IO): 30.436ms
        Processing time (w/o IO): 30.597ms
        Processing time (w/o IO): 30.448ms
        Processing time (w/o IO): 30.367ms
        Processing time (w/o IO): 30.223ms
        Processing time (w/o IO): 30.442ms
        Processing time (w/o IO): 30.69ms
          Time (mean ± σ):      56.1 ms ±   0.6 ms    [User: 45.2 ms, System: 10.9 ms]
          Range (min … max):    55.2 ms …  57.2 ms    10 runs
         
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 441.369ms
        Processing time (w/o IO): 435.063ms
        Processing time (w/o IO): 433.216ms
          Time (mean ± σ):     533.4 ms ±   1.4 ms    [User: 495.9 ms, System: 37.4 ms]
          Range (min … max):   532.4 ms … 534.4 ms    2 runs
         
Nim:

        Benchmark 1: ./related
        Processing time (w/o IO): 3825.508ms
        Processing time (w/o IO): 3820.796ms
        Processing time (w/o IO): 3817.013ms
          Time (mean ± σ):      4.170 s ±  0.003 s    [User: 4.051 s, System: 0.109 s]
          Range (min … max):    4.168 s …  4.172 s    2 runs
```