@michaelsbradleyjr 2023-10-18
Working on it.

Just curious, what Azure datacenter is your VM in?  If you don't want to say for some reason, no problem. Mine's in East US 2.